Rewrite tests to focus on our behavior
The existing tests were integrated, often exercising a full HTTP round
trip (via Rack + Rack-Test), which is fine. But they were also often
testing OmniAuth or OmniAuth-OAuth2 behavior, rather than the custom
behavior this Gem adds. They were also missing coverage for a lot of
this Gem's behavior. This is an in-place re-write of the existing test
suite to focus on the behavior of the Heroku strategy, while ignoring
(as best we can) behavior controlled at lower levels (OmniAuth or
OmniAuth-OAuth2). This also meant we could remove some dev-time
dependencies.