Remove superfluous method override
As of 3794d51b1030118785c294dd21d8586ba96c0744 we depend on the behavior
of OmniAuth-OAuth2 1.7+ which will evaluate options given as blocks for
us. This means by the time we get to this code, the `scope` is no longer
a block, and doesn't respond to `#call`; we'll always `super` and then
do nothing. Meaning we no longer need this override.